pyproject.toml
[project]
name = "tms"
version = "0.1.0"
description = "Task Management System"
requires-python = ">=3.8"
dependencies = [
    "fastmcp==2.2.7",
    "sqlalchemy==2.0.40",
    "pyyaml==6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto" 